audiopolicy: return proper devices from getDevicesForStream
HDMI/Display Port does not behave like a fixed volume
device during playback of compressed streams. This is
because getDeviceForStream(AUDIO_STREAM_MUSIC) returns
HDMI + Speaker which is not treated as a fixed volume device.

getDeviceForStream returns HDMI + Speaker as getDeviceForStrategy
returns Speaker for STRATEGY_ACCESSIBILITY which is appended to
the list although the stream is not active.

Add only those devices which have outputs with the current
stream active to fix the issue.

CRs-Fixed: 2014016
Change-Id: Ia7bf565ed51011da040564c6628f67a671355013